requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
bower==0.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-safedelete==1.0.0
django-soft-delete==0.9.18
django-vanilla-views==3.0.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
Markdown==3.3.4
mysqlclient==2.0.3
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil==1.5
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.63.0
urllib3==1.26.6
virtualenv==20.7.0
virtualenv-clone==0.5.6
whitenoise==5.3.0
